repositioning nav bar
